fix bug with mod operator

+++ playground/storage_exmpl/buffer.h
@@ -1,6 +1,8 @@
 #ifndef BUFFER_H
 #define BUFFER_H
 
+#include <string.h>
+#include <assert.h>
 #include <stdlib.h>
 #include "iodata.h"
 
@@ -9,8 +11,8 @@
 typedef struct {
 	io_data data[MAX_BUFFER_SIZE];
 	size_t size;
-	size_t head;
-	size_t tail;
+	int head;
+	int tail;
 } buffer;
 
 void init(buffer *d);
